+++ src/node_blf.h
@@ -42,7 +42,7 @@
 #define u_int64_t uint64_t
 #endif
 
-#ifdef _WIN32 
+#ifdef _WIN32
 #define u_int8_t unsigned __int8
 #define u_int16_t unsigned __int16
 #define u_int32_t unsigned __int32
@@ -103,9 +103,9 @@ void blf_cbc_decrypt(blf_ctx *, u_int8_t *, u_int8_t *, u_int32_t);
 u_int32_t Blowfish_stream2word(const u_int8_t *, u_int16_t , u_int16_t *);
 
 /* bcrypt functions*/
-void bcrypt_gensalt(u_int8_t, u_int8_t*, char *);
+void bcrypt_gensalt(char, u_int8_t, u_int8_t*, char *);
 void bcrypt(const char *, const char *, char *);
-void encode_salt(char *, u_int8_t *, u_int16_t, u_int8_t);
+void encode_salt(char *, u_int8_t *, char, u_int16_t, u_int8_t);
 u_int32_t bcrypt_get_rounds(const char *);
 
 #endif